+++ packages/examples/src/common/model.ts
@@ -1,0 +1,15 @@
+/* --------------------------------------------------------------------------------------------
+ * Copyright (c) 2018-2022 TypeFox GmbH (http://www.typefox.io). All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ * ------------------------------------------------------------------------------------------ */
+import * as cp from 'child_process';
+import { ServerOptions } from 'ws';
+export interface LanguageServerRunConfig {
+    serverName: string;
+    pathName: string;
+    serverPort: number;
+    runCommand: string; // 'node', 'java' , etc.
+    runCommandArgs: string[];
+    wsServerOptions: ServerOptions,
+    spawnOptions?: cp.SpawnOptions;
+}
