+++ packages/examples/src/index.ts
@@ -2,9 +2,8 @@
  * Copyright (c) 2024 TypeFox GmbH (http://www.typefox.io). All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-
+/* client side export only */
 export * from './common/client-commons.js';
-export * from './common/language-server-runner.js';
 export * from './browser/main.js';
 export * from './json/client/main.js';
 export * from './langium/statemachineClient.js';
