+++ packages/examples/src/node.ts
@@ -2,8 +2,9 @@
  * Copyright (c) 2024 TypeFox GmbH (http://www.typefox.io). All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-
+/* server side export only */
 export * from './common/server-commons.js';
+export * from './common/language-server-runner.js'
 export * from './json/server/json-server.js';
 export * from './json/server/main.js';
 export * from './python/server/main.js';
